src/web/django/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-crispy-forms==1.9.1
django-leaflet==0.26.0
django-range-slider==0.2.4
elasticsearch==7.7.1
idna==2.9
pytz==2020.1
requests==2.23.0
requests-aws4auth==0.9
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9